Dato for samlivsbrudd bruker ny kontekst
