+++ src/søknader/felles/steg/1-omdeg/sivilstatus/begrunnelse/DatoForSamlivsbrudd.tsx
@@ -1,29 +1,35 @@
 import React from 'react';
 import {
-  Datovelger,
   DatoBegrensning,
+  Datovelger,
 } from '../../../../../../components/dato/Datovelger';
 import LocaleTekst from '../../../../../../language/LocaleTekst';
 import KomponentGruppe from '../../../../../../components/gruppe/KomponentGruppe';
-import { IDatoFelt } from '../../../../../../models/søknad/søknadsfelter';
 import AlertStripeDokumentasjon from '../../../../../../components/AlertstripeDokumentasjon';
+import { useOmDeg } from '../../OmDegContext';
+import { useLokalIntlContext } from '../../../../../../context/LokalIntlContext';
 
-interface Props {
-  settDato: (date: string, objektnøkkel: string, tekstid: string) => void;
-  datoForSamlivsbrudd: IDatoFelt | undefined;
-}
-
-const DatoForSamlivsbrudd: React.FC<Props> = ({
-  settDato,
-  datoForSamlivsbrudd,
-}) => {
+const DatoForSamlivsbrudd: React.FC = () => {
+  const { sivilstatus, settSivilstatus } = useOmDeg();
+  const { datoForSamlivsbrudd } = sivilstatus;
   const datovelgerLabel = 'sivilstatus.datovelger.samlivsbrudd';
+  const intl = useLokalIntlContext();
+
+  const settDatoForSamlivsbrudd = (date: string, tekstid: string): void => {
+    settSivilstatus({
+      ...sivilstatus,
+      datoForSamlivsbrudd: {
+        label: intl.formatMessage({ id: tekstid }),
+        verdi: date,
+      },
+    });
+  };
 
   return (
     <>
       <KomponentGruppe>
         <Datovelger
-          settDato={(e) => settDato(e, 'datoForSamlivsbrudd', datovelgerLabel)}
+          settDato={(e) => settDatoForSamlivsbrudd(e, datovelgerLabel)}
           valgtDato={datoForSamlivsbrudd ? datoForSamlivsbrudd?.verdi : ''}
           tekstid={datovelgerLabel}
           datobegrensning={DatoBegrensning.TidligereDatoer}
